pyproject.toml
[tool.poetry]
name = "pydraw"
version = "0.1.0"
description = "A simple app to draw and animate various letters and words"
authors = ["Dhananjay <d.max1998@gmail.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.9"
black = "^21.7b0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
